requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
coverage==7.7.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.0.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
